pyproject.toml
[project]
name = "profile-to-peft"
version = "0.0.1"
authors = [{ name = "Zhaoxuan Tan" }]
description = "The code repository for the paper 'Instant Personalized Large Language Model Adaptation via Hypernetwork"
readme = "README.md"
requires-python = ">= 3.10"
dependencies = [
    "transformers",
    "huggingface_hub[cli]",
    "peft",
    "accelerate",
    "datasets",
    "bitsandbytes",
    "jupyter",
    "matplotlib",
    "hf_transfer",
    "torchmetrics",
    "inflect",
    "zstandard",
    "pre-commit",
    "trl",
    "tensorboardX",
    "wandb",
    "fasttext-wheel",
    "rouge-score",
    "evalplus",
    "gradio>=5.29.0",
    "vllm==0.5.4",
    "gpustat",
    "scikit-learn",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.isort]
profile = "black"


[tool.black]
line-length = 119

[tool.pyright]
exclude = ["train_outputs", "data", "outputs", "plots", "tmp", "wandb"]
typeCheckingMode = "off"

[tool.uv.sources]
evalplus = { git = "https://github.com/evalplus/evalplus", rev = "1895d2f6aa8895044a7cf69defc24bd57695e885" }
